Generate email to editors for a new package submission

Generate email to editors for a new package submission

Usage

slack_editors(
  package_review_issue_url = "https://github.com/ropensci/software-review/issues/000",
  package_repo_url = "https://github.com/jhondoe/coolr",
  type = "standard",
  category = "category 'workflow automation'",
  package = "coolr",
  package_does = "does cool things in R",
  package_check_shows = "this package is in top shape",
  in_scope_because = "it streamlines necessary processes",
  my_checks_show = "the package is awesome",
  my_checks_url =
    "https://github.com/ropensci/software-review/issues/000#issuecomment-0000000000",
  category_description =
    "Tools that automate and link together workflows, such as build systems and tools to manage continuous integration. Does not include general tools for literate programming. (e.g., R markdown extensions not under the previous topics). (Example: drake)"
)

app_slack_editors()

Arguments

package_review_issue_url

URL of the package review issue.

package_repo_url

URL of the package repository.

type

Type of submission (e.g., "standard").

category

Category of the package (e.g., "workflow automation").

package

Name of the package.

package_does

What the package does.

package_check_shows

Summary of the package check results.

in_scope_because

Reason why the package is in scope.

my_checks_show

Summary of the preliminary checks performed.

my_checks_url

URL to the detailed preliminary checks.

category_description

Description of the category.

Value

Character. The generated email content.

Examples

slack_editors()
#> Dear @editors we have a new standard submission under the category 'workflow automation' (description below).
#> 
#> 'coolr' does cool things in R. I believe it's in scope because it streamlines necessary processes.
#> 
#> 'pkgcheck' shows this package is in top shape. And my own quick [preliminary checks](my_checks_url) suggest the package is awesome.
#> 
#> If you can take it please double-check it's in scope and of acceptable quality and feel free to disagree with my quick preliminary checks. Then please go ahead and assign yourself with `@ropensci-review-bot assign @username as editor`.
#> 
#> Else I'll come back here after a few days.
#> 
#> ---
#> 
#> - Issue: https://github.com/ropensci/software-review/issues/000
#> - Repo: https://github.com/jhondoe/coolr
#> 
#> > category 'workflow automation': Tools that automate and link together workflows, such as build systems and tools to manage continuous integration. Does not include general tools for literate programming. (e.g., R markdown extensions not under the previous topics). (Example: drake) -- https://devguide.ropensci.org/softwarereview_policies.html#aims-and-scope